pyproject.toml
[project]
name = "last-AWS"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asgiref==3.10.0",
    "attrs==25.4.0",
    "black==25.9.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.1.8",
    "django==5.2.7",
    "django-allauth>=65.13.0",
    "django-environ>=0.12.0",
    "django-filter==25.2",
    "django-redis>=6.0.0",
    "django-restframework==0.0.1",
    "django-stubs>=5.2.7",
    "django-stubs-ext>=5.2.7",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt[blacklist]==5.5.1",
    "dotenv==0.9.9",
    "drf-spectacular==0.28.0",
    "drf-yasg==1.21.11",
    "generics>=7.0.0",
    "google-auth==2.42.0",
    "google-genai==1.47.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "inflection==0.5.1",
    "isort==7.0.0",
    "itsdangerous==2.2.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "mypy==1.18.2",
    "mypy-extensions==1.1.0",
    "openai>=2.6.1",
    "packaging==25.0",
    "pathspec==0.12.1",
    "platformdirs==4.5.0",
    "psycopg-binary==3.2.12",
    "psycopg2-binary==2.9.11",
    "psycopg[binary]==3.2.12",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pyjwt==2.10.1",
    "python-dotenv==1.2.1",
    "python-json-logger==4.0.0",
    "pytokens==0.2.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis>=7.0.1",
    "referencing==0.37.0",
    "requests==2.32.5",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "ruff==0.14.2",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "tenacity==9.1.2",
    "types-requests>=2.32.4.20250913",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "websockets==15.0.1",
    "whitenoise==6.11.0",
]

[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["migrations", "__pycache__", ".venv"]

[dependency-groups]
dev = [
    "django-stubs>=5.2.7",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict_optional = true
explicit_package_bases = true
namespace_packages = true
mypy_path = ["."]

[tool.django-stubs]
django_settings_module = "apps.settings"
strict_settings = false
apps = [
    "django.contrib.admin",
    "django.contrib.auth",
    "django.contrib.contenttypes",
    "django.contrib.sessions",
    "django.contrib.messages",
    "django.contrib.staticfiles",
    "apps.users",
    "rest_framework",
    "rest_framework_simplejwt",
    "drf_spectacular",
    "drf_yasg",
]
